refactor: Cleanup redundant code
