+++ src/components/tooltips/APRTooltip/APRTooltip.vue
@@ -6,12 +6,12 @@ import { isVeBalPool, totalAprLabel } from '@/composables/usePool';
 import { APR_THRESHOLD } from '@/constants/pools';
 import { bnum } from '@/lib/utils';
 import { Pool } from '@/services/pool/types';
-import { hasStakingRewards } from '@/services/staking/utils';
 
 import StakingBreakdown from './components/StakingBreakdown.vue';
 import VeBalBreakdown from './components/VeBalBreakdown.vue';
 import YieldBreakdown from './components/YieldBreakdown.vue';
 import { AprBreakdown } from '@balancer-labs/sdk';
+import { hasStakingRewards } from '@/composables/useAPR';
 
 /**
  * TYPES
